Moves Utils file to separate utils package and updates references
