+++ photo-portfolio/src/main/java/com/throughapinhole/photoportfolio/utils/Utils.java
@@ -1,11 +1,11 @@
-package com.throughapinhole.photoportfolio.entities;
+package com.throughapinhole.photoportfolio.utils;
 
 /**
  * Contains constant values and utility methods for use in the project
  */
-final class Utils {
+final public class Utils {
     /** Value for a Category which has no parent */
-    final static int CATEGORY_NO_PARENT = -1; 
+    final public static int CATEGORY_NO_PARENT = -1; 
 
     private Utils() {}
 }
